majit: spell the int-field width default as i64, and update the layout tuple's test callers
The macro's width for an undeclared integer field was `size_of::<usize>()`,
which is 4 on wasm32. `scalar_size` reports `size_of::<i64>()` for every
non-`Ref` field, so the default spells that instead; a `Ref` field's width comes
from `Type::Ref` and ignores the declared one either way.

`generate_inline_helper_jitcode_with_calls` took a new `int_fields` parameter
and the registered layout tuple grew two elements, but six `#[cfg(test)]` call
sites in majit-macros and majit-metainterp still used the old arity. `cargo
check` does not compile those, so only `cargo test` sees them. Also reformat the
`newlist_clear` layout literal, which `cargo fmt --check` rejected.

+++ majit/majit-macros/src/jit_interp/jitcode_lower/lower_vable.rs
@@ -23,8 +23,11 @@ pub(super) fn field_scalar_tokens(
                 const _: fn(&#struct_path) -> #ty = |__s| __s.#member;
             },
         ),
+        // `scalar_size`'s own default for a non-`Ref` field: the `i64` storage,
+        // which is 8 on every target and NOT the machine word (a `usize` here
+        // would report 4 on wasm32).
         None => (
-            quote! { ::core::mem::size_of::<usize>() },
+            quote! { ::core::mem::size_of::<i64>() },
             quote! { true },
             quote! {},
         ),
